+++ R/summary.R
@@ -7,6 +7,7 @@
 #' @param object An object of class `big_sp_list`.
 #' @param best.only Whether to return only one row corresponding to the best
 #'   model? The best model is the one smallest `$validation_loss`.
+#' @param sort Whether to sort by `$validation_loss`. Default is `FALSE`.
 #' @param ... Not used.
 #'
 #' @return A tibble with, for each `$alpha`, a mean `$validation_loss`, a mean
@@ -17,7 +18,7 @@
 #' @import foreach
 #' @importFrom tibble tibble
 #'
-summary.big_sp_list <- function(object, best.only = FALSE, ...) {
+summary.big_sp_list <- function(object, best.only = FALSE, sort = FALSE, ...) {
 
   assert_nodots()
 
@@ -28,11 +29,18 @@ summary.big_sp_list <- function(object, best.only = FALSE, ...) {
       intercept = mean(sapply(mods, function(mod) mod$intercept)),
       beta = list(rowMeans(sapply(mods, function(mod) mod$beta))),
       nb_var = sapply(beta, function(x) sum(x != 0)),
-      message = list(sapply(mods, function(mod) mod$message))
+      message = list(sapply(mods, function(mod) mod$message)),
+      all_conv = sapply(message, function(msg) all(msg == "No more improvement"))
     )
   }
 
-  `if`(best.only, res[which.min(res$validation_loss), ], res)
+  if (best.only) {
+    res[which.min(res$validation_loss), ]
+  } else if (sort) {
+    res[order(res$validation_loss, -res$alpha), ]
+  } else {
+    res
+  }
 }
 
 ################################################################################