+++ tests/testthat/test-spLinReg.R
@@ -34,7 +34,8 @@ if (not_cran) {
 
       for (covar in sample(lcovar, 1)) {
 
-        expect_warning(big_spLinReg(X, y2, covar.train = covar, ncores = test_cores()),
+        expect_warning(big_spLinReg(X, y2, covar.train = covar,
+                                    ncores = test_cores(), warn = FALSE),
                        "'y.train' is composed of only two different levels.", fixed = TRUE)
         expect_error(big_spLinReg(X, y3, covar.train = covar, ncores = test_cores()),
                      "'y.train' should be composed of different values.", fixed = TRUE)
@@ -45,7 +46,7 @@ if (not_cran) {
         alphas <- c(runif(1, min = 0.01, max = 1), 1)
 
         mod.bigstatsr <- big_spLinReg(X, y, covar.train = covar, alphas = alphas,
-                                      dfmax = Inf, n.abort = Inf,
+                                      dfmax = Inf, n.abort = Inf, warn = FALSE,
                                       ncores = test_cores())
         lapply(mod.bigstatsr, function(mod) lapply(mod, function(fold) {
           expect_length(fold$lambda, 200); NULL
@@ -66,7 +67,8 @@ if (not_cran) {
         mod.bigstatsr2 <- big_spLinReg(X, y[ind], ind.train = ind,
                                        covar.train = covar[ind, ],
                                        alphas = alphas,
-                                       ncores = test_cores())
+                                       ncores = test_cores(),
+                                       warn = FALSE)
         preds2 <- predict(mod.bigstatsr2, X, ind.row = (1:N)[-ind],
                           covar.row = covar[-ind, ])
         expect_gt(cor(preds2, y[-ind]), 0.6)
@@ -108,7 +110,8 @@ if (not_cran) {
                                          ind.col = ind.col,
                                          covar.train = covar[ind, ],
                                          alphas = alpha,
-                                         ncores = test_cores()),
+                                         ncores = test_cores(),
+                                         warn = FALSE),
           "10 variables with low/no variation have been removed.", fixed = TRUE
         )
         preds3 <- predict(mod.bigstatsr3, X, ind.row = (1:N)[-ind],
@@ -120,7 +123,8 @@ if (not_cran) {
           mod.bigstatsr4 <- big_spLinReg(X, y[ind], ind.train = ind,
                                          covar.train = covar[ind, ],
                                          alphas = alpha,
-                                         ncores = test_cores()),
+                                         ncores = test_cores(),
+                                         warn = FALSE),
           "10 variables with low/no variation have been removed.", fixed = TRUE
         )
         expect_equal(length(attr(mod.bigstatsr4, "ind.col")), M - 10)
@@ -134,7 +138,8 @@ if (not_cran) {
           mod.bigstatsr5 <- big_spLinReg(X, y[ind], ind.train = ind,
                                          covar.train = covar[ind, ],
                                          alphas = alpha, pf.X = pf,
-                                         ncores = test_cores()),
+                                         ncores = test_cores(),
+                                         warn = FALSE),
           "10 variables with low/no variation have been removed.", fixed = TRUE
         )
         preds5 <- predict(mod.bigstatsr5, X, ind.row = (1:N)[-ind],
@@ -145,7 +150,8 @@ if (not_cran) {
           mod.bigstatsr6 <- big_spLinReg(X, y[ind], ind.train = ind,
                                          covar.train = covar[ind, ],
                                          alphas = alpha, pf.X = pf,
-                                         ncores = test_cores()),
+                                         ncores = test_cores(),
+                                         warn = FALSE),
           "10 variables with low/no variation have been removed.", fixed = TRUE
         )
         lapply(unlist(mod.bigstatsr6, recursive = FALSE),
@@ -173,10 +179,10 @@ if (not_cran) {
 
         expect_warning(
           big_spLinReg(X, y, covar.train = covar, alphas = alpha,
-                       return.all = TRUE, ncores = test_cores()))
+                       return.all = TRUE, warn = FALSE, ncores = test_cores()))
 
         mod.bigstatsr4 <- big_spLinReg(X, y, covar.train = covar, alphas = alpha,
-                                       ncores = test_cores())
+                                       ncores = test_cores(), warn = FALSE)
 
         expect_length(mod.bigstatsr4, 1)
         flatten <- mod.bigstatsr4[[1]]
@@ -214,7 +220,8 @@ if (not_cran) {
                                       covar.train = covar[ind, ],
                                       alphas = alphas,
                                       ind.sets = ind.sets,
-                                      ncores = test_cores())
+                                      ncores = test_cores(),
+                                      warn = FALSE)
         preds <- predict(mod.bigstatsr, X, covar.row = covar)
         expect_gt(cor(preds[-ind], y[-ind]), 0.6)
 
@@ -225,7 +232,8 @@ if (not_cran) {
                                        covar.train = covar[ind, ],
                                        alphas = alphas,
                                        ind.sets = ind.sets,
-                                       ncores = test_cores())
+                                       ncores = test_cores(),
+                                       warn = FALSE)
         expect_equal(sapply(unlist(mod.bigstatsr2, recursive = FALSE),
                             function(x) x$intercept) + 10,
                      sapply(unlist(mod.bigstatsr, recursive = FALSE),
@@ -240,7 +248,8 @@ if (not_cran) {
                                        covar.train = covar[ind, ],
                                        alphas = alphas,
                                        ind.sets = ind.sets,
-                                       ncores = test_cores())
+                                       ncores = test_cores(),
+                                       warn = FALSE)
         expect_error(predict(mod.bigstatsr3, X, covar.row = covar),
                      "You forgot to provide 'base.row' in predict().")
         preds3 <- predict(mod.bigstatsr3, X, covar.row = covar, base.row = rep(0, N))
@@ -250,5 +259,40 @@ if (not_cran) {
   })
 
 }
+
+################################################################################
+
+test_that("Warns if not all converged", {
+  skip_if_not(not_cran)
+  set.seed(1)
+
+  # simulating some data
+  N <- 230
+  M <- 730
+  X <- FBM(N, M, init = rnorm(N * M, sd = 5))
+  y <- rowSums(X[, 1:10]) + rnorm(N)
+  covar <- matrix(rnorm(N * 3), N)
+
+  ind.train <- sort(sample(nrow(X), 150))
+  ind.test <- setdiff(rows_along(X), ind.train)
+
+  # fitting model for multiple lambdas and alphas
+  ALPHAS <- c(1, 0.5, 0.1, 0.01)
+  expect_warning(
+    test <- big_spLinReg(X, y[ind.train], ind.train = ind.train,
+                         covar.train = covar[ind.train, ],
+                         alphas = ALPHAS, K = 5),
+    "Some models may not have reached a minimum", fixed = TRUE
+  )
+
+  # peek at the models
+  test_summary <- summary(test)
+  expect_identical(test_summary$alpha, sort(ALPHAS))
+  expect_identical(test_summary$all_conv, c(FALSE, FALSE, FALSE, TRUE))
+  test_summary2 <- summary(test, sort = TRUE)
+  expect_identical(test_summary2$alpha, ALPHAS)
+  expect_identical(test_summary2$all_conv, c(TRUE, FALSE, FALSE, FALSE))
+  expect_false(is.unsorted(test_summary2$validation_loss))
+})
 
 ################################################################################